+++ CHANGELOG.md
@@ -12,6 +12,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Changed
+- If the browser uses a dark theme, the default theme is also set to dark
+
+
 ## [0.4.0] - 2023-09-26
 
 ### Added